Created Globals and Helpers
Created a Global Resource VI to hold globally accessible Message Queues and Events so they can be used throughout the Framework.  Helper methods to give these values, and to release them when the Framework is finished.  Added helper method to Parameters to help create them and a change to Read Command as part of its development.