Added Means to Inject Commands During a Test
It is now possible to inject new Commands (Command Directors) into the Test Processor whilst the test is actually running.  Added the necessary events and message types to do this and made some adjustments to the example app to add an inject case in the MPL.  The Instrument Identification app does not inject commands however.