Moved away from Test to Interaction.  Created Readme
Created a Readme document to describe how to use the framework.
There have been a few minor changes to classes.
The main change is to move away from the idea of this framework being for "testing" and to "interacting".  Initially, I had envisaged the framework being used to run tests against projects being developed.  I've now realised that is an unnecessarily limited view as the framework is generic so can be used for any purpose that requires interacting with instruments.